android/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx
sqlalchemy
psycopg2-binary
openai
python-multipart
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
httpx==0.25.2
pydantic==2.5.0